+++ app/Http/Controllers/Admin/KelasController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Jurusan;
 use App\Models\Kelas;
+use App\Models\Siswa;
 use Illuminate\Http\Request;
 
 class KelasController extends Controller
@@ -18,9 +19,9 @@ class KelasController extends Controller
         return view('admin.kelolakelas.tambah');
     }
 
-    public function store(Request $request){
+    public function store(Request $request, Kelas $kelas){
         $data = $request->validate([
-            'kelas' => 'required|string|max:255',
+            'kelas' => 'required|string|max:255|unique:kelas,kelas,' . $kelas->id,
         ]);
 
         Kelas::create([
@@ -36,7 +37,7 @@ class KelasController extends Controller
 
     public function update(Request $request, Kelas $kela){
         $data = $request->validate([
-            'kelas' => 'required|string|max:255' . $kela->id, 
+            'kelas' => 'required|string|max:255|unique:kelas,kelas,' . $kela->id, 
         ]);
 
         $kela->update([
@@ -47,6 +48,10 @@ class KelasController extends Controller
     }
 
     public function destroy(Kelas $kela){
+        $hasSiswa = Siswa::where('id_kelas', $kela->id)->exists();
+        if ($hasSiswa) {
+            return redirect()->route('admin.kelas.index')->with('error', 'Kelas tidak dapat dihapus karena masih masih memiliki siswa.');
+        }
         $kela->delete();
         return redirect()->route('admin.kelas.index')->with('success', 'kelas berhasil dihapus');
     }
